requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Werkzeug==3.1.3
Jinja2==3.1.6